+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [0.2.9]
 
-- Changed: Default runtime tag is now 'latest-stable', error check improved with image usage
+- Added: Visual progress bar when downloading image from Docker registry
+- Changed: Default runtime tag is now 'latest', secondary tag 'dev' for cincan tools, error check improved with image usage
 - Changed: registry reworked - lives in separate module cincan-registry
 - Added: With change of registry module, list subcommand has many new features, such as listing image sizes, versions and more
 - Added: With change of registry module, feature for tool version check on runtime enabled